requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
strawberry-graphql[fastapi]==0.219.0
strawberry-graphql[apollo-federation]==0.219.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Machine Learning - Scikit Learn
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Deep Learning - TensorFlow
tensorflow==2.15.0

# NLP (para clasificación de texto)
nltk==3.8.1

# Time Series Forecasting
statsmodels==0.14.1
prophet==1.1.5

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

